- fixed battery saving bug when save type equals Flash or Sram - initialize sram and flash memory to 0xff instead of 0x00 - clean flash/sram on load/reload - fixed a compilation warning

+++ src/win32/MainWndFile.cpp
@@ -728,7 +728,7 @@ void MainWnd::OnFileSavegameOldestslot()
   CString name;
   CFileStatus status;
   CString str;
-  unsigned long time = -1;
+  unsigned long time = (unsigned long)-1;
   int found = 0;
     
   for(int i = 0; i < 10; i++) {
